Create README.md
Source: https://stackoverflow.com/questions/6453879/how-to-log-quickfix-message-in-human-readable-format#:~:text=5%20Answers&text=There%20is%20no%20method%20in,to%20a%20file%20or%20DB.

Depending on the user's need there are three solutions presented here.

1. Marked-up XML
2. List of fields (groups embedded).
3. JSON-like output.

In the quickfix Python library, the FieldMap field and group key iterators are not exposed. So the approach is to first generate the XML and iterate over the tree. There is also no access to the getFieldType method of the DataDictionary, so the dictionary must be pre-processed to store the field types for conversion and handling of groups.